further ngpvanapi-client refactoring
